requirements.txt
fastapi[standard]
langchain-core
langchain-redis
langchain-ollama
langchain-groq
redis
aio-pika
sqlalchemy
psycopg2
python-dotenv